scripts/python/requirements.txt
backports.functools-lru-cache==1.5
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==2.2.2
numpy==1.14.4
pandas==0.23.0
pathlib==1.0.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
scipy==1.1.0
seaborn==0.8.1
six==1.11.0
subprocess32==3.5.2
